add additional exception if cannot get tag
