+++ django_revision/__init__.py
@@ -1,2 +1,3 @@
+from .check_revision import check_revision
 from .revision import Revision, site_revision
 from .revision_field import RevisionField
