--- django_revision/models.py
@@ -1,4 +1,0 @@
-from django.conf import settings
-
-if settings.APP_NAME == "django_revision":
-    from .tests import models  # noqa
