+++ runtests.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 import logging
-import os
 import sys
 from os.path import abspath, dirname
