+++ setup.py
@@ -19,15 +19,13 @@ setup(
     author_email="ew2789@gmail.com",
     packages=find_packages(),
     include_package_data=True,
-    url='http://github.com/erikvw/django-revision',
-    license='GPL license, see LICENSE',
-    description='Track the git revision with every model instance saved.',
+    url="https://github.com/erikvw/django-revision",
+    license="GPL license, see LICENSE",
+    description="Track the git revision with every model instance saved.",
     long_description=README,
     zip_safe=False,
     keywords="django fields git",
-    install_requires=[
-        "GitPython",
-    ],
+    install_requires=["GitPython"],
     classifiers=[
         "Environment :: Web Environment",
         "Framework :: Django",
