+++ tox.ini
@@ -1,17 +1,14 @@
 [tox]
 envlist =
-    py{37,38,39}-dj{31,32,dev}-{mysql},
+    py{39}-dj{32,dev}-{mysql},
     lint
 
 [gh-actions]
 python =
-    3.7: py37
-    3.8: py38, lint
-    3.9: py39
+    3.9: py39, lint
 
 [gh-actions:env]
 DJANGO =
-    3.1: dj31
     3.2: dj32
     dev: djdev
 
@@ -26,19 +23,19 @@ deps =
     -r requirements.txt
     -r https://raw.githubusercontent.com/clinicedc/edc/develop/requirements.tests/coverage.txt
     -r https://raw.githubusercontent.com/clinicedc/edc/develop/requirements.tests/tox.txt
-    dj31: Django>=3.1,<3.2
     dj32: Django>=3.2,<3.3
-    djdev: https://github.com/django/django/tarball/master
+    djdev: https://github.com/django/django/tarball/main
     mysql: -r https://raw.githubusercontent.com/clinicedc/edc/develop/requirements.tests/mysql.txt
 
 commands =
     pip install -U pip
+    pip install edc-test-utils
     mysql: coverage run -a runtests.py --database=mysql
     coverage report
 
 [testenv:lint]
 deps = -r https://raw.githubusercontent.com/clinicedc/edc/develop/requirements.tests/lint.txt
 commands =
-    isort --check --diff django_revision runtests.py setup.py --skip django_revision/migrations/
+    isort --profile=black --check --diff django_revision runtests.py setup.py --skip django_revision/migrations/
     black --check --diff django_revision runtests.py setup.py
     flake8 django_revision
